+++ apps/desktop/src/renderer/screens/main/components/NewLayout/TaskTabs/index.ts
@@ -1,0 +1,2 @@
+export { TaskTabs } from "./TaskTabs";
+export type { TaskTabsProps, WorktreeWithTask } from "./types";
