+++ apps/desktop/src/renderer/screens/main/components/NewLayout/TaskTabs/utils.ts
@@ -1,0 +1,15 @@
+import type { TaskStatus } from "../StatusIndicator";
+
+export const getStatusLabel = (status: TaskStatus): string => {
+	const labels: Record<TaskStatus, string> = {
+		planning: "Planning",
+		working: "Working",
+		"needs-feedback": "Needs Feedback",
+		"ready-to-merge": "Ready to Merge",
+		backlog: "Backlog",
+		todo: "Todo",
+		completed: "Completed",
+		canceled: "Canceled",
+	};
+	return labels[status] || "";
+};
